modify event controller

+++ app/controllers/events_controller.rb
@@ -46,7 +46,9 @@ class EventsController < ApplicationController
     event = Event.find(params[:id])
     if event.user_id == current_user.id
       event.destroy
-      File.delete("/app/app/assets/images/#{event.title}-#{current_user.id}.jpg")
+      if event.image != "noimage.jpg"
+        File.delete("/app/app/assets/images/#{event.title}-#{current_user.id}.jpg")
+      end
     end
   end
